Fix constant-memory streaming not working for `ZipDataByteString`.
This was due to space leaks:

The code to write the `-- central directory` referred to several
pure variables that contained thunks keeping any `ZipDataByteString`
alive all the way to streaming the end of the archive.

This change is the minimal fix that makes the test pass that was
added in the previous commit.